+++ source/extensions/omni.ai.langchain.core/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.0.8] - 2025-09-22
+
+### Changed
+- Added pydantic dependency
+- Removed jiter-0.10.0 dependency
+
 ## [2.0.7] - 2025-09-09
 
 ### Changed